requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
pyyaml>=6.0
sentencepiece>=0.1.99
optuna>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
evaluate>=0.4.0

